Merged PR 25: Merge jrios/346-DEN-add-git-repo-scheduling-config to dev
This branch contains the front end interface for setting the configuration for monitoring git repositories. It also contains two minor updates to the abstract scheduler.

**Configuration**
- The user will find the menu item to manage Monitor Git Configuration from the Application page's action menu. This should appear on applications who's repository type is GIT.
- The configuration can be enabled and disabled.
- The type of frequency is a required field.
- There are three frequency options for scheduling the monitor job. (DAILY,WEEKLY, RECURRING)
Daily: Hours, minutes, time period are required fields. Hour values are between 1-12 inclusive. Minute values are 0-59
inclusive.The job should run at the set time, every day.
Weekly: Day, Hours, minutes, time period are required fields. Hour values are between 1-12 inclusive. Minute values are
0-59 inclusive. This job should run weekly at the time and on the day specified.
Recurring: Hours, minutes are required fields. Hour values are between 0-23 inclusive. Minute values are 0-59
inclusive. If hours >0 and minutes = 0, the job runs every X hours at the time of setting the config. If hours = 0 and
minutes > 0, the job runs every y minutes. If hours > 0 and minutes > 0 then the job will run every x hours starting the
on the next hour and every y minutes starting on the next minutes.

**Abstract scheduler**
The first is an update to properly set the Day of the week on for jobs running on a weekly schedule. Cron requires this to be the abbreviated day name. The second adds logic to handle scheduling recurring jobs.